feat: Implement wallet connect
