+++ apps/web/src/components/ConnectWalletButton.tsx
@@ -1,14 +1,7 @@
-import { useTranslation } from '@pancakeswap/localization'
-import { WalletModalV2 } from '@pancakeswap/ui-wallets'
 import { Button, ButtonProps, FlexGap, WalletFilledV2Icon } from '@pancakeswap/uikit'
-import { createWallets, getDocLink } from 'config/wallet'
-import { useActiveChainId } from 'hooks/useActiveChainId'
-import useAuth from 'hooks/useAuth'
 
-import { ChainId } from '@pancakeswap/chains'
-import { useMemo, useState } from 'react'
-import { logGTMWalletConnectEvent } from 'utils/customGTMEventTracking'
-import { useConnect } from 'wagmi'
+import { useCallback, useState } from 'react'
+import WalletModalManager from 'components/WalletModalManager'
 import Trans from './Trans'
 
 interface ConnectWalletButtonProps extends ButtonProps {
@@ -16,18 +9,8 @@ interface ConnectWalletButtonProps extends ButtonProps {
 }
 
 const ConnectWalletButton = ({ children, withIcon, ...props }: ConnectWalletButtonProps) => {
-  const { login } = useAuth()
-  const {
-    t,
-    currentLanguage: { code },
-  } = useTranslation()
-  const { connectAsync } = useConnect()
-  const { chainId } = useActiveChainId()
   const [open, setOpen] = useState(false)
-
-  const docLink = useMemo(() => getDocLink(code), [code])
-
-  const wallets = useMemo(() => createWallets(chainId || ChainId.BSC, connectAsync), [chainId, connectAsync])
+  const handleOnDismiss = useCallback(() => setOpen(false), [])
 
   return (
     <>
@@ -42,15 +25,7 @@ const ConnectWalletButton = ({ children, withIcon, ...props }: ConnectWalletButt
           z-index: 99;
         }
       `}</style>
-      <WalletModalV2
-        docText={t('Learn How to Connect')}
-        docLink={docLink}
-        isOpen={open}
-        wallets={wallets}
-        login={login}
-        onDismiss={() => setOpen(false)}
-        onWalletConnectCallBack={logGTMWalletConnectEvent}
-      />
+      <WalletModalManager isOpen={open} onDismiss={handleOnDismiss} />
     </>
   )
 }
